Move storage key to locale util file
